+++ operator/src/main/java/oracle/kubernetes/operator/helpers/NamespaceHelper.java
@@ -10,10 +10,6 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import com.google.common.base.Strings;
-import oracle.kubernetes.operator.Main;
-import oracle.kubernetes.operator.TuningParameters;
-import oracle.kubernetes.operator.logging.LoggingFacade;
-import oracle.kubernetes.operator.logging.LoggingFactory;
 
 import static oracle.kubernetes.operator.helpers.HelmAccess.getHelmVariable;
 
@@ -21,10 +17,7 @@ import static oracle.kubernetes.operator.helpers.HelmAccess.getHelmVariable;
  * Operations for dealing with namespaces.
  */
 public class NamespaceHelper {
-  private static final LoggingFacade LOGGER = LoggingFactory.getLogger("Operator", "Operator");
-
   public static final String DEFAULT_NAMESPACE = "default";
-  public static final String SELECTION_STRATEGY_KEY = "domainNamespaceSelectionStrategy";
 
   private static final String operatorNamespace = computeOperatorNamespace();
 
@@ -48,32 +41,6 @@ public class NamespaceHelper {
           .collect(Collectors.toUnmodifiableList());
 
     return namespaces.isEmpty() ? Collections.singletonList(operatorNamespace) : namespaces;
-  }
-
-  public static TuningParameters tuningAndConfig() {
-    return TuningParameters.getInstance();
-  }
-
-  /**
-   * Gets the configured domain namespace selection strategy.
-   * @return Selection strategy
-   */
-  public static Main.Namespaces.SelectionStrategy getSelectionStrategy() {
-    Main.Namespaces.SelectionStrategy strategy =
-        Optional.ofNullable(tuningAndConfig().get(SELECTION_STRATEGY_KEY))
-              .map(Main.Namespaces.SelectionStrategy::valueOf)
-              .orElse(Main.Namespaces.SelectionStrategy.List);
-
-    if (Main.Namespaces.SelectionStrategy.List.equals(strategy) && isDeprecatedDedicated()) {
-      return Main.Namespaces.SelectionStrategy.Dedicated;
-    }
-    return strategy;
-  }
-
-  // Returns true if the deprecated way to specify the dedicated namespace strategy is being used.
-  // This value will only be used if the 'list' namespace strategy is specified or defaulted.
-  private static boolean isDeprecatedDedicated() {
-    return "true".equalsIgnoreCase(Optional.ofNullable(tuningAndConfig().get("dedicated")).orElse("false"));
   }
 
 }
